test(client): extract change track order test from the playlistTracksTest
